Former-commit-id: bf784f98411b184a98a2682bfb9b69b16e2f9e71

+++ Sources/Internal/Platform/Process.h
@@ -54,7 +54,7 @@ public:
     const FilePath& GetPath() const;
     const Vector<String>& GetArgs() const;
 
-    int GetExitCode() const;
+    long GetExitCode() const;
     
 private:
     
@@ -81,7 +81,7 @@ private:
     int pipes[2];
 #endif
 
-    int exitCode;
+    long exitCode;
     
 };
 };
